+++ packages/drupal-kit/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @pantheon-systems/drupal-kit
 
+## 1.1.1
+
+### Patch Changes
+
+- Re-publishing after a CI error
+
+## 1.1.0
+
+### Minor Changes
+
+- 64b63dc: [drupal-kit][next-drupal-starter] Update DrupalState version
+
 ## 1.0.2
 
 ### Patch Changes